requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
neo4j-graphrag==1.11.0
openai==2.6.1
cohere==5.20.0
google-genai==1.56.0
pydantic==2.11.7
qdrant-client==1.15.1
pydantic-settings==2.12.0
langchain-text-splitters==0.3.10
langchain-huggingface==0.3.1
langchain-community==0.3.29
langchain==0.3.27
fastembed==0.7.4
